refactor tests
This breaks search tests into their own classes
